@@ -1,6 +1,6 @@
 {
   "name": "parse-dashboard",
-  "version": "4.2.0",
+  "version": "5.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/ParsePlatform/parse-dashboard"
@@ -35,29 +35,29 @@
     "LICENSE"
   ],
   "dependencies": {
-    "@babel/runtime": "7.18.9",
+    "@babel/runtime": "7.19.0",
+    "@babel/runtime-corejs3": "7.19.1",
     "bcryptjs": "2.3.0",
     "body-parser": "1.20.0",
     "commander": "9.4.0",
     "connect-flash": "0.1.1",
     "cookie-session": "2.0.0",
-    "copy-to-clipboard": "3.3.1",
-    "core-js": "3.24.1",
+    "copy-to-clipboard": "3.3.2",
+    "core-js": "3.25.3",
     "csurf": "1.11.0",
     "express": "4.18.1",
-    "graphiql": "1.10.0",
-    "graphql": "16.5.0",
-    "history": "4.10.1",
+    "graphiql": "2.0.8",
+    "graphql": "16.6.0",
     "immutable": "4.1.0",
     "immutable-devtools": "0.1.5",
     "inquirer": "8.2.4",
-    "js-beautify": "1.14.4",
-    "otpauth": "8.0.1",
+    "js-beautify": "1.14.6",
+    "otpauth": "8.0.3",
     "package-json": "7.0.0",
     "parse": "3.4.2",
     "passport": "0.5.3",
     "passport-local": "1.0.0",
-    "prismjs": "1.28.0",
+    "prismjs": "1.29.0",
     "prop-types": "15.8.1",
     "qrcode": "1.5.1",
     "react": "16.14.0",
@@ -68,21 +68,19 @@
     "react-helmet": "6.1.0",
     "react-json-view": "1.21.3",
     "react-popper-tooltip": "4.4.2",
-    "react-redux": "8.0.2",
-    "react-router": "5.2.1",
-    "react-router-dom": "5.3.0",
+    "react-router-dom": "6.4.1",
     "regenerator-runtime": "0.13.9",
     "semver": "7.3.7",
-    "typescript": "4.7.4"
+    "typescript": "4.8.3"
   },
   "devDependencies": {
     "@actions/core": "1.9.1",
-    "@babel/core": "7.16.0",
-    "@babel/plugin-proposal-decorators": "7.16.4",
-    "@babel/plugin-transform-regenerator": "7.16.0",
-    "@babel/plugin-transform-runtime": "7.16.4",
-    "@babel/preset-env": "7.16.4",
-    "@babel/preset-react": "7.16.0",
+    "@babel/core": "7.19.3",
+    "@babel/eslint-parser": "7.19.1",
+    "@babel/plugin-proposal-decorators": "7.19.3",
+    "@babel/plugin-transform-runtime": "7.19.1",
+    "@babel/preset-env": "7.19.3",
+    "@babel/preset-react": "7.18.6",
     "@saithodev/semantic-release-backmerge": "2.1.2",
     "@semantic-release/changelog": "5.0.1",
     "@semantic-release/commit-analyzer": "8.0.1",
@@ -90,27 +88,25 @@
     "@semantic-release/github": "7.2.3",
     "@semantic-release/npm": "7.1.3",
     "@semantic-release/release-notes-generator": "9.0.3",
-    "@types/jest": "27.4.0",
-    "all-node-versions": "8.0.0",
-    "babel-eslint": "10.1.0",
-    "babel-loader": "8.2.3",
-    "cross-env": "7.0.2",
-    "css-loader": "6.5.1",
-    "eslint": "6.8.0",
-    "eslint-plugin-jest": "23.8.2",
-    "eslint-plugin-react": "7.19.0",
+    "@types/jest": "29.1.1",
+    "all-node-versions": "11.0.1",
+    "babel-loader": "8.2.5",
+    "css-loader": "6.7.1",
+    "eslint": "8.24.0",
+    "eslint-plugin-jest": "27.0.4",
+    "eslint-plugin-react": "7.31.8",
     "http-server": "14.0.0",
-    "jest": "24.8.0",
+    "jest": "29.1.2",
+    "jest-environment-jsdom": "29.1.2",
     "madge": "5.0.1",
     "marked": "4.0.10",
-    "node-sass": "7.0.3",
     "null-loader": "4.0.1",
-    "path-to-regexp": "3.2.0",
-    "puppeteer": "3.0.0",
+    "puppeteer": "18.0.5",
     "react-test-renderer": "16.13.1",
     "request": "2.88.2",
     "request-promise": "4.2.5",
-    "sass-loader": "12.3.0",
+    "sass": "1.55.0",
+    "sass-loader": "13.0.2",
     "semantic-release": "17.4.6",
     "style-loader": "3.3.1",
     "svg-prep": "1.0.4",
@@ -137,7 +133,7 @@
     "parse-dashboard": "./bin/parse-dashboard"
   },
   "engines": {
-    "node": ">=12.22.0"
+    "node": ">=14.20.1"
   },
   "main": "Parse-Dashboard/app.js",
   "jest": {
