@@ -25,8 +25,7 @@ async function checkNodeVersions() {
     releasedVersions,
     latestComponent: CiVersionCheck.versionComponents.minor,
     ignoreReleasedVersions: [
-      '<12.0.0', // These versions have reached their end-of-life support date
-      '>=13.0.0 <14.0.0', // These versions have reached their end-of-life support date
+      '<14.0.0', // These versions have reached their end-of-life support date
       '>=15.0.0 <16.0.0', // These versions have reached their end-of-life support date
       '>=17.0.0 <18.0.0', // These versions have reached their end-of-life support date
     ],
