@@ -15,7 +15,6 @@ import Popover        from 'components/Popover/Popover.react';
 import Position       from 'lib/Position';
 import PropTypes      from 'lib/PropTypes';
 import React          from 'react';
-import ReactDOM       from 'react-dom';
 import styles         from 'components/ChromeDatePicker/ChromeDatePicker.scss';
 
 export default class ChromeDatePicker extends React.Component {
@@ -26,10 +25,8 @@ export default class ChromeDatePicker extends React.Component {
       open: false,
       position: null,
     };
-  }
 
-  componentDidMount() {
-    this.node = ReactDOM.findDOMNode(this);
+    this.wrapRef = React.createRef()
   }
 
   toggle() {
@@ -37,9 +34,9 @@ export default class ChromeDatePicker extends React.Component {
       if (this.state.open) {
         return { open: false };
       }
-      let pos = Position.inWindow(this.node);
+      let pos = Position.inWindow(this.wrapRef.current);
       if (this.props.align === Directions.RIGHT) {
-        pos.x += this.node.clientWidth;
+        pos.x += this.wrapRef.current.clientWidth;
       }
       return {
         open: true,
@@ -92,7 +89,7 @@ export default class ChromeDatePicker extends React.Component {
     }
 
     return (
-      <div className={styles.wrap} onClick={this.toggle.bind(this)}>
+      <div className={styles.wrap} onClick={this.toggle.bind(this)} ref={this.wrapRef}>
         {content}
         {popover}
       </div>
