@@ -17,7 +17,6 @@ import Popover                          from 'components/Popover/Popover.react';
 import Position                         from 'lib/Position';
 import PropTypes                        from 'lib/PropTypes';
 import React                            from 'react';
-import ReactDOM                         from 'react-dom';
 import styles                           from 'components/DateRange/DateRange.scss';
 
 export default class DateRange extends React.Component {
@@ -32,10 +31,8 @@ export default class DateRange extends React.Component {
       start: val.start || monthsFrom(new Date(), -1),
       end: val.end || new Date()
     };
-  }
 
-  componentDidMount() {
-    this.node = ReactDOM.findDOMNode(this);
+    this.wrapRef = React.createRef();
   }
 
   toggle() {
@@ -43,9 +40,9 @@ export default class DateRange extends React.Component {
       if (this.state.open) {
         return { open: false };
       }
-      let pos = Position.inWindow(this.node);
+      let pos = Position.inWindow(this.wrapRef.current);
       if (this.props.align === Directions.RIGHT) {
-        pos.x += this.node.clientWidth;
+        pos.x += this.wrapRef.current.clientWidth;
       }
       return {
         open: true,
@@ -125,7 +122,7 @@ export default class DateRange extends React.Component {
     }
 
     return (
-      <div className={styles.wrap} onClick={this.toggle.bind(this)}>
+      <div className={styles.wrap} onClick={this.toggle.bind(this)} ref={this.wrapRef}>
         {content}
         {popover}
       </div>
