@@ -9,7 +9,6 @@ import Popover   from 'components/Popover/Popover.react';
 import PropTypes from 'lib/PropTypes';
 import Position  from 'lib/Position';
 import React     from 'react';
-import ReactDOM  from 'react-dom';
 import styles    from 'components/PushExperimentDropdown/PushExperimentDropdown.scss';
 
 export default class PushExperimentDropdown extends React.Component {
@@ -20,14 +19,12 @@ export default class PushExperimentDropdown extends React.Component {
       open: false,
       selected: false,
     };
+
+    this.dropdownRef = React.createRef();
   }
 
   componentWillMount() {
     this.styles = this.props.styles || styles;
-  }
-
-  componentDidMount() {
-    this.node = ReactDOM.findDOMNode(this);
   }
 
   select(value, e) {
@@ -50,7 +47,7 @@ export default class PushExperimentDropdown extends React.Component {
       </div>
     );
     if (this.state.open) {
-      let position = Position.inWindow(this.node);
+      let position = Position.inWindow(this.dropdownRef.current);
       content = (
         <Popover fixed={true} position={position} onExternalClick={() => this.setState({ open: false })}>
           <div style={widthStyle} className={[styles.menu, styles[color]].join(' ')}>
@@ -60,7 +57,7 @@ export default class PushExperimentDropdown extends React.Component {
       );
     }
     return (
-      <div style={widthStyle} className={styles.dropdown}>
+      <div style={widthStyle} className={styles.dropdown} ref={this.dropdownRef}>
         {content}
       </div>
     );
