@@ -8,7 +8,6 @@
 import Popover  from 'components/Popover/Popover.react';
 import React    from 'react';
 import styles   from 'components/SuggestionsList/SuggestionsList.scss';
-import Position from 'lib/Position';
 
 export default class Suggestion extends React.Component {
   constructor() {
@@ -19,17 +18,7 @@ export default class Suggestion extends React.Component {
       position: null
     };
 
-    this.popoverRef = React.createRef(null);
-
-    this.handleScroll = () => {
-      let newPosition = this.props.fixed
-        ? Position.inWindow(this.node)
-        : Position.inDocument(this.node);
-      newPosition.y += this.node.offsetHeight;
-      if (this.popoverRef.current) {
-        this.popoverRef.current.setPosition(newPosition);
-      }
-    };
+    this.popoverRef = React.createRef();
   }
 
   toggle() {
