@@ -7,14 +7,15 @@
  */
 import PropTypes       from 'lib/PropTypes';
 import { ActionTypes } from 'lib/stores/JobsStore';
-import history         from 'dashboard/history';
 import JobsForm        from 'dashboard/Data/Jobs/JobsForm.react';
 import React           from 'react';
 import subscribeTo     from 'lib/subscribeTo';
 import generatePath    from 'lib/generatePath';
 import { CurrentApp }  from 'context/currentApp';
+import { withRouter } from 'lib/withRouter';
 
 @subscribeTo('Jobs', 'jobs')
+@withRouter
 class JobEdit extends React.Component {
   static contextType = CurrentApp;
 
@@ -58,7 +59,7 @@ class JobEdit extends React.Component {
     let promise = this.props.params.jobId ?
       this.props.jobs.dispatch(ActionTypes.EDIT, { jobId: this.props.params.jobId, updates: schedule }) :
       this.props.jobs.dispatch(ActionTypes.CREATE, { schedule });
-    promise.then(() => {history.push(generatePath(this.context, 'jobs/scheduled'))});
+    promise.then(() => {this.props.navigate(generatePath(this.context, 'jobs/scheduled'))});
     return promise;
   }
 
