@@ -12,7 +12,6 @@ import Button                   from 'components/Button/Button.react';
 import DashboardView            from 'dashboard/DashboardView.react';
 import EmptyState               from 'components/EmptyState/EmptyState.react';
 import FormModal                from 'components/FormModal/FormModal.react';
-import history                  from 'dashboard/history';
 import LoaderContainer          from 'components/LoaderContainer/LoaderContainer.react';
 import Modal                    from 'components/Modal/Modal.react';
 import PushAudienceDialog       from 'components/PushAudienceDialog/PushAudienceDialog.react';
@@ -27,12 +26,13 @@ import Toolbar                  from 'components/Toolbar/Toolbar.react';
 import { formatAudienceSchema } from 'lib/PushUtils';
 import { List }                 from 'immutable';
 import generatePath from 'lib/generatePath';
+import { withRouter } from 'lib/withRouter';
 
 const XHR_KEY = 'PushAudiencesIndex';
 
-export default
 @subscribeTo('Schema', 'schema')
 @subscribeTo('PushAudiences', 'pushaudiences')
+@withRouter
 class PushAudiencesIndex extends DashboardView {
   constructor() {
     super();
@@ -111,7 +111,7 @@ class PushAudiencesIndex extends DashboardView {
   }
 
   handleSendPush(objectId) {
-    history.push(generatePath(this.context, `push/new?audienceId=${objectId}`));
+    this.props.navigate(generatePath(this.context, `push/new?audienceId=${objectId}`));
   }
 
   renderRow(audience) {
@@ -245,7 +245,7 @@ class PushAudiencesIndex extends DashboardView {
       }}>
     </FormModal>
 
-    if (typeof(data) !== undefined) {
+    if (typeof data !== 'undefined') {
       if (data.size === 0) {
         content = <div className={stylesTable.empty}>{this.renderEmpty()}</div>;
       } else {
@@ -276,3 +276,5 @@ class PushAudiencesIndex extends DashboardView {
     );
   }
 }
+
+export default PushAudiencesIndex;
