@@ -15,7 +15,6 @@ import Field                  from 'components/Field/Field.react';
 import Fieldset               from 'components/Fieldset/Fieldset.react';
 import FieldStyles            from 'components/Field/Field.scss';
 import FlowView               from 'components/FlowView/FlowView.react';
-import history                from 'dashboard/history';
 import Label                  from 'components/Label/Label.react';
 import LoaderContainer        from 'components/LoaderContainer/LoaderContainer.react';
 import Parse                  from 'parse';
@@ -33,6 +32,7 @@ import { Directions }         from 'lib/Constants';
 import { Link }               from 'react-router-dom';
 import { tableInfoBuilder }   from 'lib/PushUtils';
 import generatePath from 'lib/generatePath';
+import { withRouter } from 'lib/withRouter';
 
 const EXP_STATS_URL = 'http://docs.parseplatform.org/ios/guide/#push-experiments';
 
@@ -200,8 +200,8 @@ const COLOR_MAP = {
 const DROPDOWN_KEY_GROUP_A = 'Group A';
 const DROPDOWN_KEY_GROUP_B = 'Group B';
 
-export default
 @subscribeTo('Schema', 'schema')
+@withRouter
 class PushDetails extends DashboardView {
   constructor() {
     super();
@@ -472,7 +472,7 @@ class PushDetails extends DashboardView {
       prevLaunchGroup = (
         <div className={styles.header}>
           <div className={styles.headline}>
-            This push is the Launch Group for a previous <Link to={{ pathname: getPushDetailUrl(this.context, pushDetails.experiment_push_id) }}>experiment</Link>.
+            This push is the Launch Group for a previous <Link to={getPushDetailUrl(this.context, pushDetails.experiment_push_id)}>experiment</Link>.
           </div>
         </div>
       );
@@ -586,7 +586,7 @@ class PushDetails extends DashboardView {
       if (error) {
         promise.reject({ error });
       } else {
-        history.push(generatePath(this.context, 'push/activity'));
+        this.props.navigate(generatePath(this.context, 'push/activity'));
       }
     }, (error) => {
       promise.reject({ error });
@@ -744,3 +744,5 @@ class PushDetails extends DashboardView {
     );
   }
 }
+
+export default PushDetails;
