@@ -6,20 +6,22 @@
  * the root directory of this source tree.
  */
 import * as ComponentsMap from 'parse-interface-guide/ComponentsMap';
-import { Link }           from 'react-router-dom';
+import { NavLink }        from 'react-router-dom';
 import Icon               from 'components/Icon/Icon.react';
 import PropsTable         from 'parse-interface-guide/PropsTable.react';
 import React              from 'react';
 import styles             from 'parse-interface-guide/PIG.scss';
 import beautify           from 'js-beautify';
 import CodeSnippet        from 'components/CodeSnippet/CodeSnippet.react';
+import { withRouter }     from 'lib/withRouter';
 
 let PIGRow = ({ title, children }) => <div>
   <div className={styles.header}>{title}</div>
   <div className={styles.row}>{children}</div>
 </div>;
 
-export default class PIG extends React.Component {
+@withRouter
+class PIG extends React.Component {
   constructor() {
     super();
 
@@ -46,7 +48,7 @@ export default class PIG extends React.Component {
           }}/>
         {components.map((name) => {
           return name.toLowerCase().indexOf(this.state.query.toLowerCase()) !== -1
-            ? <Link activeClassName={styles.active} key={name} to={`/${name}`}>{name}</Link>
+            ? <NavLink className={({ isActive }) => isActive ? styles.active : undefined} key={name} to={`/${name}`}>{name}</NavLink>
             : null;
         })}
       </div>
@@ -54,7 +56,7 @@ export default class PIG extends React.Component {
   }
 
   renderContent() {
-    let componentInfo = ComponentsMap[this.props.params.component];
+    let componentInfo = ComponentsMap[this.props.params['*']];
     if (!componentInfo) {
       componentInfo = ComponentsMap[Object.keys(ComponentsMap)[0]];
     }
@@ -89,3 +91,5 @@ export default class PIG extends React.Component {
     );
   }
 }
+
+export default PIG;
