@@ -12,7 +12,9 @@ module.exports = {
   process: function (src, filename) {
     if (filename.endsWith('.scss') || filename.endsWith('.css')) {
       var matches = extractClassnames(src);
-      return 'module.exports = ' + JSON.stringify(matches);
+      return {
+        code: 'module.exports = ' + JSON.stringify(matches),
+      };
     }
 
     // Let Jest handle our custom module resolution
@@ -26,14 +28,14 @@ module.exports = {
       return babel.transform(src, {
         filename: filename,
         retainLines: true,
-        plugins: [['@babel/plugin-proposal-decorators', { 'legacy': true }], '@babel/transform-regenerator', '@babel/transform-runtime'],
-        presets: ['@babel/preset-react', '@babel/preset-env']
         // Remove propTypes for tests so we don't have to keep unmocking lib/PropTypes
         // Also it's more representative of the production environment
         //plugins: [ 'babel-plugin-remove-proptypes' ]
-      }).code;
+      });
     }
 
-    return src;
+    return {
+      code: src,
+    };
   }
 };
